Generowanie ścieżek (nie wiem czy w 100% git, sprawdzę potem)

+++ src/PunktKolizji.java
@@ -1,9 +1,9 @@
 public class PunktKolizji {
     private int sciezka1;
     private int sciezka2;
-    private double pozx;
-    private double pozy;
-    PunktKolizji (int sciezka1, int sciezka2, double pozx, double pozy) {
+    private int pozx;
+    private int pozy;
+    PunktKolizji (int sciezka1, int sciezka2, int pozx, int pozy) {
         this.sciezka1 = sciezka1;
         this.sciezka2 = sciezka2;
         this.pozx = pozx;
